Refactor chat and space selector drawer to a single widget
