+++ app/lib/common/providers/chat_providers.dart
@@ -9,9 +9,6 @@ import 'package:riverpod/riverpod.dart';
 
 final _log = Logger('a3::common::chat');
 
-final chatSearchValueProvider =
-    StateProvider.autoDispose<String?>((ref) => null);
-
 /// Provider the profile data of a the given space, keeps up to date with underlying client
 final convoProvider =
     AsyncNotifierProvider.family<AsyncConvoNotifier, Convo?, Convo>(
