+++ app/lib/common/widgets/chat/chat_selector_drawer.dart
@@ -1,10 +1,6 @@
-import 'package:acter/common/providers/chat_providers.dart';
-
-import 'package:acter/common/widgets/room/brief_room_list_entry.dart';
+import 'package:acter/common/widgets/room/select_room_drawer.dart';
 import 'package:acter_avatar/acter_avatar.dart';
-import 'package:atlas_icons/atlas_icons.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 
 const Key selectChatDrawerKey = Key('chat-widgets-select-chat-drawer');
@@ -21,56 +17,13 @@ Future<String?> selectChatDrawer({
     enableDrag: true,
     context: context,
     isDismissible: true,
-    builder: (context) => Consumer(
-      builder: (context, ref, child) {
-        final chats = ref.watch(chatsProvider);
-        return Padding(
-          padding: const EdgeInsets.all(10),
-          child: Column(
-            key: key,
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisSize: MainAxisSize.min,
-            children: <Widget>[
-              Row(
-                children: [
-                  Expanded(
-                    child: title ?? Text(L10n.of(context).selectChat),
-                  ),
-                  OutlinedButton.icon(
-                    icon: const Icon(Atlas.minus_circle_thin),
-                    onPressed: () {
-                      Navigator.pop(context, '');
-                    },
-                    label: Text(L10n.of(context).clear),
-                  ),
-                ],
-              ),
-              Flexible(
-                child: chats.isEmpty
-                    ? Text(L10n.of(context).noChatsFound)
-                    : ListView.builder(
-                        padding: const EdgeInsets.all(8),
-                        itemCount: chats.length,
-                        itemBuilder: (context, index) {
-                          final convo = chats[index];
-                          return BriefRoomEntry(
-                            roomId: convo.getRoomIdStr(),
-                            avatarDisplayMode: DisplayMode.GroupChat,
-                            keyPrefix: 'select-chat',
-                            selectedValue: currentChatId,
-                            canCheck: canCheck,
-                            onSelect: (roomId) {
-                              Navigator.pop(context, roomId);
-                            },
-                          );
-                        },
-                      ),
-              ),
-            ],
-          ),
-        );
-      },
+    builder: (context) => SelectRoomDrawer(
+      key: key,
+      canCheck: canCheck,
+      title: title ?? Text(L10n.of(context).selectChat),
+      avatarDisplayMode: DisplayMode.GroupChat,
+      keyPrefix: 'select-chat',
+      roomType: RoomType.groupChat,
     ),
   );
   if (selected == null) {
